Now all the logic to log in and sign in are done, it validates if the users exist to log in, if the user exist in the sign in, and it created the user in the sign in function if it does not exist.
Missing to add the logic to create the users directory (now it need the folder), discuss with my partner if the logic to create the user file is OK, missing the logic to validate if the user is already with a session and redirect to home.

All the redirect to home function are commented until the integration of the branches. and missing to add the "Go Back" button.